totalContributors and lastMonthContributors should work on Profiles
